@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "mxcube_video_streamer"
-version = "1.3.0"
+version = "1.4.0"
 packages = [{include = "video_streamer"}]
 license = "MIT"
 description = "FastAPI Based video streamer"
@@ -23,24 +23,25 @@ classifiers = [
 ]
 
 [tool.poetry.dependencies]
-python = ">=3.7,<3.12"
-pydantic = "^1.10.5"
-fastapi = "^0.92.0"
-pillow = "^9.4.0"
-uvicorn = "^0.20.0"
-pytango = { version = "^9.3.0", optional=true}
-jinja2 = "^3.1.2"
-websockets = "^10.4"
+python = ">=3.8,<3.12"
+pydantic = ">=2.8.2,<2.9.0"
+pydantic-settings = ">=2.4.0"
+fastapi = "^0.112.1"
+pillow = "^10.4.0"
+uvicorn = "^0.30.6"
+pytango = { version = "^9.4.1", optional=true}
+jinja2 = "^3.1.4"
+websockets = "^13.0"
 
 
 [tool.poetry.dev-dependencies]
-pylint = [ { version = "==2.13.9", python = "<=3.7.1" }, { version = "2.15.3", python = ">=3.8" } ]
-pre-commit = "2.20.0"
-pytest = "7.1.3"
-pytest-cov = "4.0.0"
+pre-commit = "3.5.0"
+pytest = "8.3.2"
+pytest-cov = "5.0.0"
 
 [tool.black]
 line-length = 88
+target_version = ["py38"]
 include = '\.pyi?$'
 exclude = '''
 /(
@@ -55,6 +56,22 @@ exclude = '''
   | dist
 )/
 '''
+
+[tool.pyright]
+ignore = ["**/typing_extensions.py"]
+pythonVersion = "3.8"
+
+[tool.ruff]
+line-length = 88
+indent-width = 4
+target-version = "py38"
+
+[tool.ruff.lint]
+extend-select = ["B"]
+
+[tool.ruff.format]
+quote-style = "double"
+indent-style = "space"
 
 [build-system]
 requires = ["poetry-core>=1.0.0", "setuptools>=45", "setuptools_scm[toml]>=6.2"]
